+++ FallRiver/FallRiver/NPC.cpp
@@ -110,7 +110,7 @@ void NPC::Update(float fElapsedTime)
 		if(pDI->KeyPressed(DIK_RETURN) || pDI->JoystickButtonPressed(0,0))
 		{
 			bool questEmpty = true;
-			pGPS->m_pHUD->SetTarget(pGPS->hospitalX,pGPS->hospitalY);
+			pGPS->m_pHUD->SetTarget((float)pGPS->hospitalX,(float)pGPS->hospitalY);
 
 
 			for(unsigned int i = 0; i < pGPS->GetPlayer()->m_vpActiveQuests.size(); i++)
@@ -221,9 +221,9 @@ void NPC::Update(float fElapsedTime)
 			if(strcmp(thisAnim->frames[m_playerAnim.curAnimation][m_playerAnim.curFrame].eventMsg,"none") != 0)
 				EventSystem::GetInstance()->SendEvent(thisAnim->frames[m_playerAnim.curAnimation][m_playerAnim.curFrame].eventMsg, this);
 		}
-		if((m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size()) && thisAnim->looping[m_playerAnim.curAnimation])
+		if(((unsigned int)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size()) && thisAnim->looping[m_playerAnim.curAnimation])
 			m_playerAnim.curFrame = 0;
-		else if(m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size() && !thisAnim->looping[m_playerAnim.curAnimation])
+		else if((unsigned int)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size() && !thisAnim->looping[m_playerAnim.curAnimation])
 			m_playerAnim.curFrame--;
 	}
 }
